Move filtering logic from fetching resource
