+++ src/examinations/+hooks/useTableDialogs.jsx
@@ -4,3 +4,8 @@ export const useVisibilityDialog = () => {
   const controls = useDialog("Visibility");
   return controls;
 };
+
+export const useExportDialog = () => {
+  const controls = useDialog("Export");
+  return controls;
+};
